+++ app/src/main/java/com/theandroiddev/mywins/remote/SuccessesRemoteImpl.kt
@@ -1,7 +1,7 @@
 package com.theandroiddev.mywins.remote
 
-import com.theandroiddev.mywins.data.model.SuccessEntity
-import com.theandroiddev.mywins.data.repository.SuccessesRemoteDataSource
+import com.theandroiddev.mywins.data.successes.model.SuccessEntity
+import com.theandroiddev.mywins.data.successes.repository.SuccessesRemoteDataSource
 import com.theandroiddev.mywins.remote.model.toEntity
 import io.reactivex.Single
 import javax.inject.Inject
